Fix URL to 'User roles API' in 1st note
The first Note card in this flow contains an invalid link to an internal 'urldefense' tool. This PR removes the reference and updates it with a link to the Okta documentation. 

This change should be tested by someone other than me for verification. Contact me if questions Pablo.valarezo@liveramp.com or @pabst in the macadmins Slack.